@@ -11,9 +11,9 @@ const { ipcRenderer } = window.require('electron') as {
 }
 import { InfoBox, ToggleSwitch, SvgButton } from 'src/components/UI'
 
-import Backspace from '@material-ui/icons/Backspace'
+import Backspace from '@mui/icons-material/Backspace'
 import ContextProvider from 'src/state/ContextProvider'
-import CreateNewFolder from '@material-ui/icons/CreateNewFolder'
+import CreateNewFolder from '@mui/icons-material/CreateNewFolder'
 
 interface Props {
   appName: string
