@@ -5,12 +5,15 @@ import { existsSync, readFileSync } from 'graceful-fs'
 
 import { execAsync, isOnline } from './utils'
 import { execOptions, heroicToolsPath, home } from './constants'
-import { logError, logInfo, logWarning } from './logger'
+import { logError, logInfo, LogPrefix, logWarning } from './logger/logger'
 
 export const DXVK = {
   getLatest: async () => {
     if (!(await isOnline())) {
-      logWarning('App offline, skipping possible DXVK update.')
+      logWarning(
+        'App offline, skipping possible DXVK update.',
+        LogPrefix.DXVKInstaller
+      )
       return
     }
 
@@ -45,19 +48,22 @@ export const DXVK = {
       const echoCommand = `echo ${pkg} > ${heroicToolsPath}/${tool.name}/latest_${tool.name}`
       const cleanCommand = `rm ${latestVersion}`
 
-      logInfo(`Updating ${tool.name} to:`, pkg)
+      logInfo([`Updating ${tool.name} to:`, pkg], LogPrefix.DXVKInstaller)
 
       return execAsync(downloadCommand)
         .then(async () => {
-          logInfo(`downloaded ${tool.name}`)
-          logInfo(`extracting ${tool.name}`)
+          logInfo(`downloaded ${tool.name}`, LogPrefix.DXVKInstaller)
+          logInfo(`extracting ${tool.name}`, LogPrefix.DXVKInstaller)
           exec(echoCommand)
           await execAsync(extractCommand)
-          logInfo(`extracting ${tool.name} updated!`)
+          logInfo(`extracting ${tool.name} updated!`, LogPrefix.DXVKInstaller)
           exec(cleanCommand)
         })
         .catch((error) =>
-          logError(`Error when downloading ${tool.name}`, error)
+          logError(
+            [`Error when downloading ${tool.name}`, error],
+            LogPrefix.DXVKInstaller
+          )
         )
     })
   },
@@ -72,7 +78,10 @@ export const DXVK = {
     const isValidPrefix = existsSync(`${winePrefix}/.update-timestamp`)
 
     if (!isValidPrefix) {
-      logWarning('DXVK cannot be installed on a Proton or a invalid prefix!')
+      logWarning(
+        'DXVK cannot be installed on a Proton or a invalid prefix!',
+        LogPrefix.DXVKInstaller
+      )
       return
     }
 
@@ -82,7 +91,7 @@ export const DXVK = {
     const wineBin = winePathSplit.join('/').concat('/')
 
     if (!existsSync(`${heroicToolsPath}/${tool}/latest_${tool}`)) {
-      logError('dxvk not found!')
+      logError('dxvk not found!', LogPrefix.DXVKInstaller)
       await DXVK.getLatest()
     }
 
@@ -105,17 +114,20 @@ export const DXVK = {
     const updatedVersionfile = `echo '${globalVersion}' > ${currentVersionCheck}`
 
     if (action === 'restore') {
-      logInfo(`Removing ${tool} version information`)
+      logInfo(`Removing ${tool} version information`, LogPrefix.DXVKInstaller)
       const updatedVersionfile = `rm -rf ${currentVersionCheck}`
       const removeCommand = `PATH=${wineBin}:$PATH WINEPREFIX='${winePrefix}' bash ${toolPath}/setup_dxvk.sh uninstall --symlink`
       return execAsync(removeCommand, execOptions)
         .then(() => {
-          logInfo(`${tool} removed from ${winePrefix}`)
+          logInfo(`${tool} removed from ${winePrefix}`, LogPrefix.DXVKInstaller)
           return exec(updatedVersionfile)
         })
         .catch((error) => {
-          logError(error)
-          logError('error when removing DXVK, please try again')
+          logError(`${error}`, LogPrefix.DXVKInstaller)
+          logError(
+            'error when removing DXVK, please try again',
+            LogPrefix.DXVKInstaller
+          )
         })
     }
 
@@ -123,16 +135,17 @@ export const DXVK = {
       return
     }
 
-    logInfo([`installing ${tool} on...`, prefix])
+    logInfo([`installing ${tool} on...`, prefix], LogPrefix.DXVKInstaller)
     await execAsync(installCommand, { shell: '/bin/bash' })
       .then(() => {
-        logInfo(`${tool} installed on ${winePrefix}`)
+        logInfo(`${tool} installed on ${winePrefix}`, LogPrefix.DXVKInstaller)
         return exec(updatedVersionfile)
       })
       .catch((error) => {
-        logError(error)
+        logError(`${error}`, LogPrefix.DXVKInstaller)
         logError(
-          'error when installing DXVK, please try launching the game again'
+          'error when installing DXVK, please try launching the game again',
+          LogPrefix.DXVKInstaller
         )
       })
   }
