@@ -16,14 +16,14 @@ import {
 } from '../types'
 import { LegendaryGame } from './games'
 import { LegendaryUser } from './user'
-import { execAsync, isEpicOffline, isOnline } from '../utils'
+import { execAsync, isEpicServiceOffline, isOnline } from '../utils'
 import {
   installed,
   legendaryBin,
   legendaryConfigPath,
   libraryPath
 } from '../constants'
-import { logError, logInfo, logWarning } from '../logger'
+import { logError, logInfo, LogPrefix, logWarning } from '../logger/logger'
 import { spawn } from 'child_process'
 import Store from 'electron-store'
 import { GlobalConfig } from '../config'
@@ -82,11 +82,14 @@ class LegendaryLibrary {
    * Refresh library.
    */
   public async refresh() {
-    logInfo('Refreshing Epic Games...')
+    logInfo('Refreshing Epic Games...', LogPrefix.Legendary)
     const { showUnrealMarket } = await GlobalConfig.get().getSettings()
-    const epicOffline = await isEpicOffline()
+    const epicOffline = await isEpicServiceOffline()
     if (epicOffline) {
-      logWarning('Epic is Offline right now, cannot update game list!')
+      logWarning(
+        'Epic is Offline right now, cannot update game list!',
+        LogPrefix.Backend
+      )
       return
     }
 
@@ -94,7 +97,11 @@ class LegendaryLibrary {
       const getUeAssets = showUnrealMarket ? '--include-ue' : ''
       const child = spawn(legendaryBin, ['list', getUeAssets])
       child.stderr.on('data', (data) => {
-        console.log(`${data}`)
+        if (`${data}`.includes('ERROR')) {
+          logError(`${data}`, LogPrefix.Legendary)
+        } else {
+          logInfo(`${data}`, LogPrefix.Legendary)
+        }
       })
       child.on('error', (err) => rej(`${err}`))
       child.on('close', () => res('finished'))
@@ -104,7 +111,7 @@ class LegendaryLibrary {
         this.loadAll()
       })
       .catch((err) => {
-        logError(err)
+        logError(err, LogPrefix.Legendary)
       })
   }
 
@@ -134,7 +141,7 @@ class LegendaryLibrary {
     format: 'info' | 'class' = 'class',
     fullRefresh?: boolean
   ): Promise<(LegendaryGame | GameInfo)[]> {
-    logInfo('Refreshing library...')
+    logInfo('Refreshing library...', LogPrefix.Legendary)
     const isLoggedIn = await LegendaryUser.isLoggedIn()
     if (!isLoggedIn) {
       return
@@ -142,10 +149,10 @@ class LegendaryLibrary {
 
     if (fullRefresh) {
       try {
-        logInfo('Legendary: Refreshing Epic Games...')
+        logInfo('Refreshing Epic Games...', LogPrefix.Legendary)
         await this.refresh()
       } catch (error) {
-        logError(error)
+        logError(`${error}`, LogPrefix.Legendary)
       }
     }
 
@@ -153,7 +160,7 @@ class LegendaryLibrary {
       this.refreshInstalled()
       await this.loadAll()
     } catch (error) {
-      logError(error)
+      logError(`${error}`, LogPrefix.Legendary)
     }
     const arr = Array.from(this.library.values()).sort(
       (a: { title: string }, b: { title: string }) => {
@@ -166,9 +173,9 @@ class LegendaryLibrary {
       if (libraryStore.has('library')) {
         libraryStore.delete('library')
       }
-      logInfo('Updating game list')
+      logInfo('Updating game list', LogPrefix.Legendary)
       libraryStore.set('library', arr)
-      logInfo('Game List Updated')
+      logInfo('Game List Updated', LogPrefix.Legendary)
       return arr
     }
     if (format === 'class') {
@@ -203,8 +210,11 @@ class LegendaryLibrary {
    */
   public async getInstallInfo(appName: string) {
     const cache = installStore.get(appName) as InstallInfo
-    const epicOffline = await isEpicOffline()
-    logInfo('Getting More details with Legendary info command... ')
+    const epicOffline = await isEpicServiceOffline()
+    logInfo(
+      'Getting More details with Legendary info command... ',
+      LogPrefix.Legendary
+    )
     if (cache) {
       return cache
     }
@@ -218,8 +228,8 @@ class LegendaryLibrary {
       installStore.set(appName, info)
       return info
     } catch (error) {
-      logError('Error running the Legendary Info command')
-      console.log({ error })
+      logError('Error running the Legendary Info command', LogPrefix.Legendary)
+      logError(`${error}`, LogPrefix.Legendary)
     }
   }
 
@@ -235,25 +245,28 @@ class LegendaryLibrary {
     if (!isLoggedIn || !online) {
       return []
     }
-    const epicOffline = await isEpicOffline()
+    const epicOffline = await isEpicServiceOffline()
     if (epicOffline) {
-      logWarning('Epic servers are offline, cannot check for game updates')
+      logWarning(
+        'Epic servers are offline, cannot check for game updates',
+        LogPrefix.Backend
+      )
       return []
     }
 
     const command = `${legendaryBin} list-installed --check-updates --tsv`
     try {
       const { stdout } = await execAsync(command)
-      logInfo('Checking for game updates')
+      logInfo('Checking for game updates', LogPrefix.Legendary)
       const updates = stdout
         .split('\n')
         .filter((item) => item.split('\t')[4] === 'True')
         .map((item) => item.split('\t')[0])
         .filter((item) => item.length > 1)
-      logInfo(`Found ${updates.length} game(s) to update`)
+      logInfo(`Found ${updates.length} game(s) to update`, LogPrefix.Legendary)
       return updates
     } catch (error) {
-      logError(error)
+      logError(`${error}`, LogPrefix.Legendary)
     }
   }
 
