@@ -3,19 +3,18 @@ import { existsSync, readFileSync } from 'graceful-fs'
 import { UserInfo } from '../types'
 import { clearCache, execAsync } from '../utils'
 import { legendaryBin, userInfo } from '../constants'
-import { logError, logInfo, LogPrefix } from '../logger'
+import { logError, logInfo, LogPrefix } from '../logger/logger'
 import { spawn } from 'child_process'
 import { userInfo as user } from 'os'
 import Store from 'electron-store'
-import { dialog, session } from 'electron'
-import i18next from 'i18next'
+import { session } from 'electron'
 
 const configStore = new Store({
   cwd: 'store'
 })
 export class LegendaryUser {
   public static async login(sid: string) {
-    logInfo('Logging with Legendary...')
+    logInfo('Logging with Legendary...', LogPrefix.Legendary)
 
     const command = `auth --sid ${sid}`.split(' ')
     return new Promise((res) => {
@@ -23,10 +22,6 @@ export class LegendaryUser {
       child.stderr.on('data', (data) => {
         if (`${data}`.includes('ERROR')) {
           logError(`${data}`, LogPrefix.Legendary)
-          dialog.showErrorBox(
-            i18next.t('box.error.login.tittle', 'Cannot Login!'),
-            `${data}`
-          )
           return res('error')
         } else {
           logInfo(`stderr: ${data}`, LogPrefix.Legendary)
@@ -36,10 +31,6 @@ export class LegendaryUser {
       child.stdout.on('data', (data) => {
         if (`${data}`.includes('ERROR')) {
           logError(`${data}`, LogPrefix.Legendary)
-          dialog.showErrorBox(
-            i18next.t('box.error.login.tittle', 'Cannot Login!'),
-            `${data}`
-          )
           return res('error')
         } else {
           logInfo(`stdout: ${data}`, LogPrefix.Legendary)
@@ -73,7 +64,7 @@ export class LegendaryUser {
     try {
       isLoggedIn = await LegendaryUser.isLoggedIn()
     } catch (error) {
-      logError(error)
+      logError(`${error}`, LogPrefix.Backend)
       configStore.delete('userInfo')
     }
     if (isLoggedIn) {
