@@ -1,3 +1,4 @@
+import { WineVersionInfo } from './wine-downloader/types'
 import {
   InstallParams,
   LaunchResult,
@@ -21,7 +22,7 @@ import { cpus, platform } from 'os'
 import {
   existsSync,
   mkdirSync,
-  rmdirSync,
+  rmSync,
   unlinkSync,
   watch,
   writeFile
@@ -36,21 +37,22 @@ import { GlobalConfig } from './config'
 import { LegendaryLibrary } from './legendary/library'
 import { LegendaryUser } from './legendary/user'
 import {
-  checkCommandVersion,
   checkForUpdates,
   clearCache,
   errorHandler,
   execAsync,
-  isEpicOffline,
+  isEpicServiceOffline,
   getLegendaryVersion,
   getSystemInfo,
   handleExit,
   isOnline,
   openUrlOrFile,
   resetHeroic,
-  showAboutWindow
+  showAboutWindow,
+  showItemInFolder
 } from './utils'
 import {
+  currentLogFile,
   discordLink,
   execOptions,
   getShell,
@@ -71,8 +73,14 @@ import {
   wikiLink
 } from './constants'
 import { handleProtocol } from './protocol'
-import { logError, logInfo, LogPrefix, logWarning } from './logger'
+import { logError, logInfo, LogPrefix, logWarning } from './logger/logger'
 import Store from 'electron-store'
+import {
+  updateWineVersionInfos,
+  removeWineVersion,
+  installWineVersion
+} from './wine-downloader/utils'
+import { ProgressInfo, State } from 'heroic-wine-downloader'
 
 const { showErrorBox, showMessageBox, showOpenDialog } = dialog
 const isWindows = platform() === 'win32'
@@ -94,12 +102,31 @@ const tsStore = new Store({
 async function createWindow(): Promise<BrowserWindow> {
   const { exitToTray, startInTray } = await GlobalConfig.get().getSettings()
 
+  let windowProps: Electron.Rectangle = {
+    height: 690,
+    width: 1200,
+    x: 0,
+    y: 0
+  }
+
+  if (store.has('window-props')) {
+    const tmpWindowProps = store.get('window-props') as Electron.Rectangle
+    if (
+      tmpWindowProps &&
+      tmpWindowProps.width &&
+      tmpWindowProps.height &&
+      tmpWindowProps.y !== undefined &&
+      tmpWindowProps.x !== undefined
+    ) {
+      windowProps = tmpWindowProps
+    }
+  }
+
   // Create the browser window.
   mainWindow = new BrowserWindow({
-    height: 690,
+    ...windowProps,
     minHeight: 650,
     minWidth: 1100,
-    width: 1200,
     show: !(exitToTray && startInTray),
     webPreferences: {
       webviewTag: true,
@@ -121,22 +148,13 @@ async function createWindow(): Promise<BrowserWindow> {
   app.setAppUserModelId('Heroic')
   app.commandLine.appendSwitch('enable-spatial-navigation')
 
-  if (!app.isPackaged) {
-    /* eslint-disable @typescript-eslint/ban-ts-comment */
-    //@ts-ignore
-    import('electron-devtools-installer').then((devtools) => {
-      const { default: installExtension, REACT_DEVELOPER_TOOLS } = devtools
-
-      installExtension(REACT_DEVELOPER_TOOLS).catch((err: string) => {
-        logError(['An error occurred: ', err])
-      })
-    })
-    mainWindow.loadURL('http://localhost:3000')
-    // Open the DevTools.
-    mainWindow.webContents.openDevTools()
-
+  const onMainWindowClose = async () => {
     mainWindow.on('close', async (e) => {
       e.preventDefault()
+
+      // store windows properties
+      store.set('window-props', mainWindow.getBounds())
+
       const { exitToTray } = await GlobalConfig.get().config
 
       if (exitToTray) {
@@ -145,18 +163,27 @@ async function createWindow(): Promise<BrowserWindow> {
 
       return await handleExit()
     })
+  }
+
+  if (!app.isPackaged) {
+    /* eslint-disable @typescript-eslint/ban-ts-comment */
+    //@ts-ignore
+    import('electron-devtools-installer').then((devtools) => {
+      const { default: installExtension, REACT_DEVELOPER_TOOLS } = devtools
+
+      installExtension(REACT_DEVELOPER_TOOLS).catch((err: string) => {
+        logError(['An error occurred: ', err], LogPrefix.Backend)
+      })
+    })
+    mainWindow.loadURL('http://localhost:3000')
+    // Open the DevTools.
+    mainWindow.webContents.openDevTools()
+
+    onMainWindowClose()
   } else {
     Menu.setApplicationMenu(null)
 
-    mainWindow.on('close', async (e) => {
-      e.preventDefault()
-      const { exitToTray } = await GlobalConfig.get().config
-
-      if (exitToTray) {
-        return mainWindow.hide()
-      }
-      return handleExit()
-    })
+    onMainWindowClose()
     mainWindow.loadURL(`file://${path.join(__dirname, '../build/index.html')}`)
 
     return mainWindow
@@ -235,13 +262,13 @@ if (!gotTheLock) {
   })
   app.whenReady().then(async () => {
     const systemInfo = await getSystemInfo()
-    logInfo(`${systemInfo}`)
+    logInfo(`${systemInfo}`, LogPrefix.Backend)
     // We can't use .config since apparently its not loaded fast enough.
     const { language, darkTrayIcon } = await GlobalConfig.get().getSettings()
     const isLoggedIn = await LegendaryUser.isLoggedIn()
 
     if (!isLoggedIn) {
-      logInfo('User Not Found, removing it from Store')
+      logInfo('User Not Found, removing it from Store', LogPrefix.Backend)
       store.delete('userinfo')
     }
 
@@ -295,12 +322,12 @@ if (!gotTheLock) {
     })
     if (!app.isDefaultProtocolClient('heroic')) {
       if (app.setAsDefaultProtocolClient('heroic')) {
-        logInfo('Registered protocol with OS.')
+        logInfo('Registered protocol with OS.', LogPrefix.Backend)
       } else {
-        logError('Failed to register protocol with OS.')
+        logError('Failed to register protocol with OS.', LogPrefix.Backend)
       }
     } else {
-      logWarning('Protocol already registered.')
+      logWarning('Protocol already registered.', LogPrefix.Backend)
     }
     if (process.argv[1]) {
       const url = process.argv[1]
@@ -313,14 +340,14 @@ if (!gotTheLock) {
     appIcon.setContextMenu(contextMenu())
     appIcon.setToolTip('Heroic')
     ipcMain.on('changeLanguage', async (event, language: string) => {
-      logInfo(['Changing Language to:', language])
+      logInfo(['Changing Language to:', language], LogPrefix.Backend)
       await i18next.changeLanguage(language)
       gameInfoStore.clear()
       appIcon.setContextMenu(contextMenu())
     })
 
     ipcMain.addListener('changeTrayColor', () => {
-      logInfo('Changing Tray icon Color...')
+      logInfo('Changing Tray icon Color...', LogPrefix.Backend)
       setTimeout(async () => {
         const { darkTrayIcon } = await GlobalConfig.get().getSettings()
         const trayIcon = darkTrayIcon ? iconDark : iconLight
@@ -328,26 +355,6 @@ if (!gotTheLock) {
         appIcon.setContextMenu(contextMenu())
       }, 500)
     })
-
-    if (process.platform === 'linux') {
-      const found = await checkCommandVersion(
-        ['python', 'python3'],
-        '3.8.0',
-        false
-      )
-
-      if (!found) {
-        dialog.showErrorBox(
-          i18next.t('box.error.python.title', 'Python Error!'),
-          i18next.t(
-            'box.error.python.message',
-            'Heroic requires Python 3.8 or newer.'
-          )
-        )
-        logError('Heroic requires Python 3.8 or newer.')
-        return app.quit()
-      }
-    }
 
     return
   })
@@ -380,7 +387,7 @@ ipcMain.on('Notify', (event, args) => {
 
 // Maybe this can help with white screens
 process.on('uncaughtException', (err) => {
-  logError(`${err.name}: ${err.message}`)
+  logError(`${err.name}: ${err.message}`, LogPrefix.Backend)
 })
 
 let powerId: number | null
@@ -435,23 +442,19 @@ ipcMain.on('openKofiPage', () => openUrlOrFile(kofiPage))
 ipcMain.on('openWikiLink', () => openUrlOrFile(wikiLink))
 ipcMain.on('openSidInfoPage', () => openUrlOrFile(sidInfoUrl))
 
-ipcMain.on('getLog', (event, appName) =>
-  openUrlOrFile(`${heroicGamesConfigPath}${appName}-lastPlay.log`)
-)
-
 ipcMain.on('removeFolder', async (e, [path, folderName]) => {
   if (path === 'default') {
     const { defaultInstallPath } = await GlobalConfig.get().getSettings()
     const path = defaultInstallPath.replaceAll("'", '')
     const folderToDelete = `${path}/${folderName}`
     return setTimeout(() => {
-      rmdirSync(folderToDelete, { recursive: true })
+      rmSync(folderToDelete, { recursive: true })
     }, 5000)
   }
 
   const folderToDelete = `${path}/${folderName}`.replaceAll("'", '')
   return setTimeout(() => {
-    rmdirSync(folderToDelete, { recursive: true })
+    rmSync(folderToDelete, { recursive: true })
   }, 2000)
 })
 
@@ -482,12 +485,14 @@ ipcMain.handle(
       winePrefix = `${protonPrefix}/pfx`
 
       logWarning(
-        'Using Winecfg and Winetricks with Proton might not work as expected.'
+        'Using Winecfg and Winetricks with Proton might not work as expected.',
+        LogPrefix.Backend
       )
       // workaround for proton since newer versions doesnt come with a wine binary anymore.
       if (!existsSync(wineBin.replaceAll("'", ''))) {
         logInfo(
-          `${wineBin} not found for this Proton version, will try using default wine`
+          `${wineBin} not found for this Proton version, will try using default wine`,
+          LogPrefix.Backend
         )
         wineBin = '/usr/bin/wine'
       }
@@ -501,12 +506,13 @@ ipcMain.handle(
       command = `WINEPREFIX='${winePrefix}' ${wineBin} '${exe}'`
     }
 
-    logInfo(['trying to run', command])
+    logInfo(['trying to run', command], LogPrefix.Backend)
     try {
       await execAsync(command, execOptions)
     } catch (error) {
       logError(
-        `Something went wrong! Check if ${tool} is available and ${wineBin} exists`
+        `Something went wrong! Check if ${tool} is available and ${wineBin} exists`,
+        LogPrefix.Backend
       )
     }
   }
@@ -518,7 +524,7 @@ ipcMain.handle('checkGameUpdates', () =>
   LegendaryLibrary.get().listUpdateableGames()
 )
 
-ipcMain.handle('getEpicGamesStatus', () => isEpicOffline())
+ipcMain.handle('getEpicGamesStatus', () => isEpicServiceOffline())
 
 // Not ready to be used safely yet.
 ipcMain.handle('updateAll', () => LegendaryLibrary.get().updateAllGames())
@@ -568,7 +574,7 @@ ipcMain.handle('getGameInfo', async (event, game) => {
     info.extra = await Game.get(game).getExtraInfo(info.namespace)
     return info
   } catch (error) {
-    logError(error)
+    logError(`${error}`, LogPrefix.Backend)
   }
 })
 
@@ -581,7 +587,7 @@ ipcMain.handle('getInstallInfo', async (event, game) => {
     const info = await Game.get(game).getInstallInfo()
     return info
   } catch (error) {
-    logError(error)
+    logError(`${error}`, LogPrefix.Backend)
     return {}
   }
 })
@@ -606,7 +612,10 @@ ipcMain.handle('readConfig', async (event, config_class) => {
     case 'user':
       return (await LegendaryUser.getUserInfo()).displayName
     default:
-      logError(`Which idiot requested '${config_class}' using readConfig?`)
+      logError(
+        `Which idiot requested '${config_class}' using readConfig?`,
+        LogPrefix.Backend
+      )
       return {}
   }
 })
@@ -636,13 +645,50 @@ ipcMain.handle('writeConfig', (event, [appName, config]) => {
 // Watch the installed games file and trigger a refresh on the installed games if something changes
 if (existsSync(installed)) {
   watch(installed, () => {
-    logInfo('Legendary: Installed game list updated')
+    logInfo('Installed game list updated', LogPrefix.Legendary)
     LegendaryLibrary.get().refreshInstalled()
   })
 }
 
 ipcMain.handle('refreshLibrary', async (e, fullRefresh) => {
   return await LegendaryLibrary.get().getGames('info', fullRefresh)
+})
+
+ipcMain.handle('refreshWineVersionInfo', async (e, fetch) => {
+  try {
+    const releases = await updateWineVersionInfos(fetch)
+    return releases
+  } catch (error) {
+    logError(String(error), LogPrefix.WineDownloader)
+    return
+  }
+})
+
+ipcMain.handle('installWineVersion', async (e, release: WineVersionInfo) => {
+  const onProgress = (state: State, progress: ProgressInfo) => {
+    e.sender.send('progressOf' + release.version, { state, progress })
+  }
+  notify({
+    title: `${release?.type} - ${release?.version}`,
+    body: i18next.t('notify.install.startInstall')
+  })
+  return installWineVersion(release, onProgress).then((res) => {
+    notify({
+      title: `${release?.type} - ${release?.version}`,
+      body: i18next.t('notify.install.finished')
+    })
+    return res
+  })
+})
+
+ipcMain.handle('removeWineVersion', async (e, release: WineVersionInfo) => {
+  return removeWineVersion(release).then((res) => {
+    notify({
+      title: `${release?.type} - ${release?.version}`,
+      body: i18next.t('notify.uninstalled')
+    })
+    return res
+  })
 })
 
 ipcMain.on('logError', (e, err) => logError(`${err}`, LogPrefix.Frontend))
@@ -676,7 +722,7 @@ ipcMain.handle(
       tsStore.set(`${game}.firstPlayed`, startPlayingDate)
     }
 
-    logInfo([`launching`, title, game])
+    logInfo([`launching`, title, game], LogPrefix.Backend)
 
     if (recentGames.length) {
       let updatedRecentGames = recentGames.filter((a) => a.appName !== game)
@@ -733,10 +779,9 @@ ipcMain.handle(
           status: 'done'
         })
 
-        writeFile(
-          `${heroicGamesConfigPath}${game}-lastPlay.log`,
-          logResult,
-          () => logInfo('Log was written')
+        const gameLogFile = `${heroicGamesConfigPath}${game}-lastPlay.log`
+        writeFile(gameLogFile, logResult, () =>
+          logInfo(`Log was written to ${gameLogFile}`, LogPrefix.Backend)
         )
         return stderr
       })
@@ -748,7 +793,7 @@ ipcMain.handle(
           stderr,
           () => 'done'
         )
-        logError(stderr)
+        logError(stderr, LogPrefix.Backend)
         window.webContents.send('setGameStatus', {
           appName,
           status: 'done'
@@ -766,6 +811,10 @@ ipcMain.handle('openDialog', async (e, args) => {
     return { path: filePaths[0] }
   }
   return { canceled }
+})
+
+ipcMain.on('showItemInFolder', async (e, item) => {
+  showItemInFolder(item)
 })
 
 const openMessageBox = async (args: Electron.MessageBoxOptions) => {
@@ -795,11 +844,14 @@ ipcMain.handle('install', async (event, params) => {
     platform() === 'darwin' && is_mac_native ? 'Mac' : 'Windows'
 
   if (!(await isOnline())) {
-    logWarning(`App offline, skipping install for game '${title}'.`)
+    logWarning(
+      `App offline, skipping install for game '${title}'.`,
+      LogPrefix.Backend
+    )
     return
   }
 
-  const epicOffline = await isEpicOffline()
+  const epicOffline = await isEpicServiceOffline()
   if (epicOffline) {
     dialog.showErrorBox(
       i18next.t('box.warning.title', 'Warning'),
@@ -831,7 +883,7 @@ ipcMain.handle('install', async (event, params) => {
             ? i18next.t('notify.install.finished')
             : i18next.t('notify.install.canceled')
       })
-      logInfo('finished installing')
+      logInfo('finished installing', LogPrefix.Backend)
       mainWindow.webContents.send('setGameStatus', {
         appName,
         status: 'done'
@@ -855,14 +907,17 @@ ipcMain.handle('uninstall', async (event, game) => {
     .uninstall()
     .then(() => {
       notify({ title, body: i18next.t('notify.uninstalled') })
-      logInfo('finished uninstalling')
+      logInfo('finished uninstalling', LogPrefix.Backend)
     })
-    .catch(logError)
+    .catch((error) => logError(`${error}`, LogPrefix.Backend))
 })
 
 ipcMain.handle('repair', async (event, game) => {
   if (!(await isOnline())) {
-    logWarning(`App offline, skipping repair for game '${game}'.`)
+    logWarning(
+      `App offline, skipping repair for game '${game}'.`,
+      LogPrefix.Backend
+    )
     return
   }
   const title = (await Game.get(game).getGameInfo()).title
@@ -871,14 +926,14 @@ ipcMain.handle('repair', async (event, game) => {
     .repair()
     .then(() => {
       notify({ title, body: i18next.t('notify.finished.reparing') })
-      logInfo('finished repairing')
+      logInfo('finished repairing', LogPrefix.Backend)
     })
     .catch((error) => {
       notify({
         title,
         body: i18next.t('notify.error.reparing', 'Error Repairing')
       })
-      logError(error)
+      logError(`${error}`, LogPrefix.Backend)
     })
 })
 
@@ -887,18 +942,18 @@ ipcMain.handle('moveInstall', async (event, [appName, path]: string[]) => {
   try {
     const newPath = await Game.get(appName).moveInstall(path)
     notify({ title, body: i18next.t('notify.moved') })
-    logInfo(`Finished moving ${appName} to ${newPath}.`)
+    logInfo(`Finished moving ${appName} to ${newPath}.`, LogPrefix.Backend)
   } catch (error) {
     notify({
       title,
       body: i18next.t('notify.error.move', 'Error Moving the Game')
     })
-    logError(error)
+    logError(`${error}`, LogPrefix.Backend)
   }
 })
 
 ipcMain.handle('importGame', async (event, args) => {
-  const epicOffline = await isEpicOffline()
+  const epicOffline = await isEpicServiceOffline()
   if (epicOffline) {
     dialog.showErrorBox(
       i18next.t('box.warning.title', 'Warning'),
@@ -926,7 +981,7 @@ ipcMain.handle('importGame', async (event, args) => {
         appName,
         status: 'done'
       })
-      logInfo(`imported ${title}`)
+      logInfo(`imported ${title}`, LogPrefix.Backend)
     })
     .catch((err) => {
       notify({ title, body: i18next.t('notify.install.canceled') })
@@ -934,17 +989,20 @@ ipcMain.handle('importGame', async (event, args) => {
         appName,
         status: 'done'
       })
-      logInfo(err)
+      logInfo(err, LogPrefix.Backend)
     })
 })
 
 ipcMain.handle('updateGame', async (e, game) => {
   if (!(await isOnline())) {
-    logWarning(`App offline, skipping install for game '${game}'.`)
+    logWarning(
+      `App offline, skipping install for game '${game}'.`,
+      LogPrefix.Backend
+    )
     return
   }
 
-  const epicOffline = await isEpicOffline()
+  const epicOffline = await isEpicServiceOffline()
   if (epicOffline) {
     dialog.showErrorBox(
       i18next.t('box.warning.title', 'Warning'),
@@ -969,10 +1027,10 @@ ipcMain.handle('updateGame', async (e, game) => {
             ? i18next.t('notify.update.finished')
             : i18next.t('notify.update.canceled')
       })
-      logInfo('finished updating')
+      logInfo('finished updating', LogPrefix.Backend)
     })
     .catch((err) => {
-      logError(err)
+      logError(err, LogPrefix.Backend)
       notify({ title, body: i18next.t('notify.update.canceled') })
       return err
     })
@@ -1022,7 +1080,8 @@ ipcMain.handle('requestGameProgress', async (event, appName) => {
 
   const progress = { bytes, eta, percent }
   logInfo(
-    `Progress: ${appName} ${progress.percent}/${progress.bytes}/${progress.eta}`
+    `Progress: ${appName} ${progress.percent}/${progress.bytes}/${progress.eta}`,
+    LogPrefix.Backend
   )
   return progress
 })
@@ -1031,7 +1090,7 @@ ipcMain.handle(
   'changeInstallPath',
   async (event, [appName, newPath]: string[]) => {
     LegendaryLibrary.get().changeGameInstallPath(appName, newPath)
-    logInfo(`Finished moving ${appName} to ${newPath}.`)
+    logInfo(`Finished moving ${appName} to ${newPath}.`, LogPrefix.Backend)
   }
 )
 
@@ -1056,10 +1115,14 @@ ipcMain.handle('egsSync', async (event, args) => {
     const { stderr, stdout } = await execAsync(
       `${legendaryBin} egl-sync ${command} -y`
     )
-    logInfo(`${stdout}`)
-    logError(`${stderr}`)
+    logInfo(`${stdout}`, LogPrefix.Legendary)
+    if (stderr.includes('ERROR')) {
+      logError(`${stderr}`, LogPrefix.Legendary)
+      return 'Error'
+    }
     return `${stdout} - ${stderr}`
   } catch (error) {
+    logError(`${error}`, LogPrefix.Legendary)
     return 'Error'
   }
 })
@@ -1084,19 +1147,25 @@ ipcMain.on('removeShortcut', async (event, appName: string) => {
 
 ipcMain.handle('syncSaves', async (event, args) => {
   const [arg = '', path, appName] = args
-  const epicOffline = await isEpicOffline()
+  const epicOffline = await isEpicServiceOffline()
   if (epicOffline) {
     logWarning('Epic is Offline right now, cannot sync saves!')
     return 'Epic is Offline right now, cannot sync saves!'
   }
   if (!(await isOnline())) {
-    logWarning(`App offline, skipping syncing saves for game '${appName}'.`)
+    logWarning(
+      `App offline, skipping syncing saves for game '${appName}'.`,
+      LogPrefix.Backend
+    )
     return
   }
 
   const { stderr, stdout } = await Game.get(appName).syncSaves(arg, path)
-  logInfo(`${stdout}`)
-  logError(`${stderr}`)
+  logInfo(`${stdout}`, LogPrefix.Backend)
+  if (stderr.includes('ERROR')) {
+    logError(`${stderr}`, LogPrefix.Backend)
+    return `Something went wrong, check ${currentLogFile}!`
+  }
   return `\n ${stdout} - ${stderr}`
 })
 
@@ -1200,3 +1269,8 @@ ipcMain.handle('gamepadAction', async (event, args) => {
     inputEvents.forEach((event) => window.webContents.sendInputEvent(event))
   }
 })
+
+/*
+ * INSERT OTHER IPC HANLDER HERE
+ */
+import './logger/ipc_handler'
