@@ -1,6 +1,11 @@
 import React, { PureComponent } from 'react'
 
-import { GameInfo, GameStatus, RefreshOptions } from 'src/types'
+import {
+  GameInfo,
+  GameStatus,
+  RefreshOptions,
+  WineVersionInfo
+} from 'src/types'
 import { TFunction, withTranslation } from 'react-i18next'
 import { getLegendaryConfig, getPlatform, install, launch } from 'src/helpers'
 import { i18n } from 'i18next'
@@ -19,6 +24,10 @@ const libraryStore: ElectronStore = new Store({
   cwd: 'lib-cache',
   name: 'library'
 })
+const wineDownloaderInfoStore: ElectronStore = new Store({
+  cwd: 'store',
+  name: 'wine-downloader-info'
+})
 
 const RTL_LANGUAGES = ['fa']
 
@@ -33,6 +42,7 @@ interface Props {
 interface StateProps {
   category: string
   data: GameInfo[]
+  wineVersions: WineVersionInfo[]
   error: boolean
   filter: string
   filterText: string
@@ -50,6 +60,9 @@ export class GlobalState extends PureComponent<Props> {
     category: 'games',
     data: libraryStore.has('library')
       ? (libraryStore.get('library') as GameInfo[])
+      : [],
+    wineVersions: wineDownloaderInfoStore.has('wine-releases')
+      ? (wineDownloaderInfoStore.get('wine-releases') as WineVersionInfo[])
       : [],
     error: false,
     filter: 'all',
@@ -113,6 +126,26 @@ export class GlobalState extends PureComponent<Props> {
       ipcRenderer.send('logError', error)
     }
     this.refresh(checkForUpdates)
+  }
+
+  refreshWineVersionInfo = async (fetch: boolean): Promise<void> => {
+    if (this.state.platform !== 'linux') {
+      return
+    }
+    ipcRenderer.send('logInfo', 'Refreshing wine downloader releases')
+    try {
+      this.setState({ refreshing: true })
+      const releases = await ipcRenderer.invoke('refreshWineVersionInfo', fetch)
+      this.setState({
+        wineVersions: releases,
+        refreshing: false
+      })
+    } catch (error) {
+      this.setState({ refreshing: false })
+      console.error(error)
+      ipcRenderer.send('logError', error)
+      ipcRenderer.send('logError', 'Refreshing wine downloader releases failed')
+    }
   }
 
   handleSearch = (input: string) => this.setState({ filterText: input })
@@ -412,7 +445,8 @@ export class GlobalState extends PureComponent<Props> {
 
   render() {
     const { children } = this.props
-    const { data, filterText, filter, platform, filterPlatform } = this.state
+    const { data, wineVersions, filterText, filter, platform, filterPlatform } =
+      this.state
     let filteredLibrary = data
     const language = storage.getItem('language') || 'en'
     const isRTL = RTL_LANGUAGES.includes(language)
@@ -434,6 +468,7 @@ export class GlobalState extends PureComponent<Props> {
         value={{
           ...this.state,
           data: filteredLibrary,
+          wineVersions: wineVersions,
           handleCategory: this.handleCategory,
           handleFilter: this.handleFilter,
           handleGameStatus: this.handleGameStatus,
@@ -444,6 +479,7 @@ export class GlobalState extends PureComponent<Props> {
           platform: platform,
           refresh: this.refresh,
           refreshLibrary: this.refreshLibrary,
+          refreshWineVersionInfo: this.refreshWineVersionInfo,
           recentGames: getRecentGames(data)
         }}
       >
